refactor: ループラッパー共通関数を scripts/lib/loop-runner.ts に抽出
pad2, formatTimestamp, tee, extractAssistantText の4関数が
auto-triage.ts と character-audit.ts で一字一句重複していたため、
共通モジュールに抽出して import に置き換えた。

Closes #800
